+++ Engine/Core/Core.vcxproj.filters
@@ -141,7 +141,7 @@
     <ClCompile Include="src\yougine\managers\GameManager.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="src\yougine\Editor\ProjectWindow\ProjectWindow.cpp">
+    <ClCompile Include="src\yougine\Editor\ProjectWindows\ProjectWindow.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
     <ClCompile Include="src\yougine\Project\Project.cpp">
@@ -245,7 +245,7 @@
     <ClInclude Include="src\yougine\managers\IManager.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="src\yougine\Editor\ProjectWindow\ProjectWindow.h">
+    <ClInclude Include="src\yougine\Editor\ProjectWindows\ProjectWindow.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
     <ClInclude Include="src\yougine\Project\Project.h">
