+++ Engine/Core/src/yougine/main.cpp
@@ -15,7 +15,7 @@
 #include <fstream>
 
 #include "components/TransformComponent.h"
-#include "Editor/ProjectWindow/ProjectWindow.h"
+#include "Editor/ProjectWindows/ProjectWindow.h"
 #include "managers/ComponentList.h"
 #include "managers/GameManager.h"
 #include "Project/Project.h"
@@ -27,7 +27,7 @@ static void glfw_error_callback(int error, const char* description)
 
 int main()
 {
-    auto project = Projects::Project::GetInstance();
+    auto project = projects::Project::GetInstance();
     project->projectFolderPath = "D:/Yougin/YouginProject/TestProject/script/";
     glfwSetErrorCallback(glfw_error_callback);
 
@@ -96,7 +96,7 @@ int main()
     editor_windows_manager->AddRenderWindow(new editor::HierarchyWindow(editor_windows_manager, scene, input_manager));
     editor_windows_manager->AddRenderWindow(new editor::SceneWindow(editor_windows_manager, scene));
     editor_windows_manager->AddRenderWindow(new editor::InspectorWindow(editor_windows_manager, scene, input_manager));
-    editor_windows_manager->AddRenderWindow(new ProjectWindow(editor_windows_manager,scene));
+    editor_windows_manager->AddRenderWindow(new editor::projectwindows::ProjectWindow(editor_windows_manager,scene));
 
     //GameManagerで回すマネージャのvector
     std::vector<IManager> managerlist;
